Kaggle_Serverstal_SteelDefectDetection/requirements.txt
imageio
segmentation-models-pytorch
pytorch
torchvision
torchsummary